Move AtomicWriter out of TiltFile.
This is so that we can use AtomicWriter for other files int the future.
