+++ Assets/Scripts/DevOptions.cs
@@ -43,9 +43,6 @@ using Newtonsoft.Json.Linq;
 
 namespace TiltBrush {
 
-[Serializable]
-public enum TiltFormat { Directory, Inherit, Zip };
-
 // This class awakens before all our other script classes, so it is
 // safe for it to override values that have been serialized into the scene.
 public class DevOptions : MonoBehaviour {
@@ -57,7 +54,7 @@ public class DevOptions : MonoBehaviour {
   public bool ResaveLegacyScenes = true;
   public bool UseAutoProfiler = false;
   public bool AllowStripBreak = true;
-  public TiltFormat PreferredTiltFormat = TiltFormat.Inherit;
+  public AtomicWriter.SaveFormat PreferredTiltFormat = AtomicWriter.SaveFormat.Inherit;
   public PointerScript.BrushLerp BrushLerp = PointerScript.BrushLerp.Default;
 
   public void Awake() {
