+++ Assets/Scripts/Save/AtomicWriter.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: c0c0407257cac7444bd4f25e378bb9fe
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
